do not fail on missing files or folders

+++ bin/docker-install-tutorials.sh
@@ -6,16 +6,23 @@ galaxy_instance="http://localhost:8080"
 # launch the instance
 echo " - Launch a lite instance"
 startup_lite
-sleep 120 #galaxy-wait --timeout 60
+
+# wait until galaxy has started
+galaxy-wait -g $galaxy_instance
 
 # install other tutorial materials
 for dir in /tutorials/*
 do
-    echo "For tutorial: $dir"
+    echo "Installing tutorial: $dir"
 
     # install tools
-    echo " - Installing tools"
-    shed-install -t $dir/tools.yaml -g $galaxy_instance -u $GALAXY_DEFAULT_ADMIN_USER -p $GALAXY_DEFAULT_ADMIN_PASSWORD
+    if [ -f $dir/tools.yaml ]
+    then
+        echo " - Installing tools"
+        shed-install -t $dir/tools.yaml -g $galaxy_instance -u $GALAXY_DEFAULT_ADMIN_USER -p $GALAXY_DEFAULT_ADMIN_PASSWORD
+    else
+        echo " - No tools to install for this tutorial (no file named tools.yaml present)"
+    fi
 
     # install workflows (TODO: make them shared instead of just under admin user account?)
     if [ -d $dir/workflows/ ];
@@ -23,12 +30,17 @@ do
         echo " - Installing workflows"
         workflow-install --workflow_path $dir/workflows/ -g $galaxy_instance -u $GALAXY_DEFAULT_ADMIN_USER -p $GALAXY_DEFAULT_ADMIN_PASSWORD
     else
-        echo "No workflow to install"
+        echo " - No workflows to install (no folder named workflows present)"
     fi
 
     # install data libraries
-    echo " - Installing data libraries"
-    setup-data-libraries -i $dir/data-library.yaml -g $galaxy_instance -u $GALAXY_DEFAULT_ADMIN_USER -p $GALAXY_DEFAULT_ADMIN_PASSWORD
+    if [ -f $dir/data-library.yaml ]
+    then
+        echo " - Installing data libraries"
+        setup-data-libraries -i $dir/data-library.yaml -g $galaxy_instance -u $GALAXY_DEFAULT_ADMIN_USER -p $GALAXY_DEFAULT_ADMIN_PASSWORD
+    else
+        echo " - No data libraries to install (no data-library.yaml file present)"
+    fi
 
     # install reference data? (discussion: do this at build or run time?)
     if [ -f $dir/data-manager.yaml ]
@@ -36,17 +48,25 @@ do
         echo " - Installing reference data"
         run-data-managers --config $dir/data-manager.yaml -g $galaxy_instance -u $GALAXY_DEFAULT_ADMIN_USER -p $GALAXY_DEFAULT_ADMIN_PASSWORD
     else
-        echo "No reference data to install (no data-manager.yaml file present)"
+        echo " - No reference data to install (no data-manager.yaml file present)"
     fi
 
     # install tours
     dir_name="$(dirname $dir)"
     dir_name="$(basename $dir_name)"
-    for t in $dir/tours/*
-    do
-        # prefix tour file name with tutorial name to avoid clashes
-        fname=$dir_name-$(basename $t)
-        echo " - Installing tour: $t as $fname"
-        cp $t $GALAXY_ROOT/config/plugins/tours/$fname
-    done
+    if [ -d $dir/tours/ ];
+    then
+        echo " - Installing tours"
+        for t in $dir/tours/*
+        do
+            # prefix tour file name with tutorial name to avoid clashes
+            fname=$dir_name-$(basename $t)
+            echo " - Installing tour: $t as $fname"
+            cp $t $GALAXY_ROOT/config/plugins/tours/$fname
+        done
+    else
+        echo " - No tours to install for this tutorial (no directory named tours present)"
+    fi
+
+    echo "Finished installation of tutorial \n"
 done
